+++ src/helpers/prepare-cargo-media-source.js
@@ -1,7 +1,10 @@
-const prepareCargoMediaSource = ({ src, imgWidth}) => {
-  const width = (imgWidth * 2).toFixed();
+const prepareCargoMediaSource = ({ src, imgWidth, originalImgWidth }) => {
+  const width =
+    imgWidth !== originalImgWidth && imgWidth * 2 < originalImgWidth
+      ? imgWidth * 2
+      : originalImgWidth;
 
-  return src.replace('/t/original/', `/w/${width}/q/75/`);
+  return src.replace('/t/original/', `/w/${width.toFixed()}/q/75/`);
 };
 
 export default prepareCargoMediaSource;